Feat: Fix all deletion issues, add graceful shutdown to everything, new delete confirmation process

+++ app/Livewire/Project/Service/ServiceApplicationView.php
@@ -4,12 +4,18 @@ namespace App\Livewire\Project\Service;
 
 use App\Models\ServiceApplication;
 use Livewire\Component;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
 
 class ServiceApplicationView extends Component
 {
     public ServiceApplication $application;
 
     public $parameters;
+
+    public $docker_cleanup = true;
+    public $delete_volumes = true;
+ 
 
     protected $rules = [
         'application.human_name' => 'nullable',
@@ -23,12 +29,7 @@ class ServiceApplicationView extends Component
         'application.is_stripprefix_enabled' => 'nullable|boolean',
     ];
 
-    public function render()
-    {
-        return view('livewire.project.service.service-application-view');
-    }
-
-    public function updatedApplicationFqdn()
+       public function updatedApplicationFqdn()
     {
         $this->application->fqdn = str($this->application->fqdn)->replaceEnd(',', '')->trim();
         $this->application->fqdn = str($this->application->fqdn)->replaceStart(',', '')->trim();
@@ -56,8 +57,13 @@ class ServiceApplicationView extends Component
         $this->dispatch('success', 'You need to restart the service for the changes to take effect.');
     }
 
-    public function delete()
+    public function delete($password)
     {
+        if (!Hash::check($password, Auth::user()->password)) {
+            $this->addError('password', 'The provided password is incorrect.');
+            return;
+        }
+
         try {
             $this->application->delete();
             $this->dispatch('success', 'Application deleted.');
@@ -90,5 +96,18 @@ class ServiceApplicationView extends Component
         } finally {
             $this->dispatch('generateDockerCompose');
         }
+    }
+
+    public function render()
+    {
+        return view('livewire.project.service.service-application-view', [
+            'checkboxes' => [
+                ['id' => 'delete_volumes', 'label' => 'All associated volumes with this resource will be permanently deleted'],
+                ['id' => 'docker_cleanup', 'label' => 'Docker cleanup will be run on the server which removes builder cache and unused images'],
+                // ['id' => 'delete_associated_backups_locally', 'label' => 'All backups associated with this Ressource will be permanently deleted from local storage.'],
+                // ['id' => 'delete_associated_backups_s3', 'label' => 'All backups associated with this Ressource will be permanently deleted from the selected S3 Storage.'],
+                // ['id' => 'delete_associated_backups_sftp', 'label' => 'All backups associated with this Ressource will be permanently deleted from the selected SFTP Storage.']
+            ]
+        ]);
     }
 }
